pyproject.toml
[project]
name = "stego-llm"
version = "0.1.0"
description = "LLM Steganography - Embed and extract hidden messages in LLM-generated text"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-cpp-python>=0.3.16",
    "numpy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["stego_llm"]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pytest>=8.4.1",
]
